pyproject.toml
[project]
name = "table-maker"
version = "0.1.0"
description = "Generate random tables as PNG images and JSON data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=36.2.2",
    "pillow>=11.1.0",
]

[project.scripts]
table-maker = "table_maker.cli:main"

[dependency-groups]
dev = [
    "pyright>=1.1.396",
    "ruff>=0.9.9",
    "pytest>=8.0.0",
]

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
